renamed navigation buttons to reflect their use
